chore: tackle tech debt with inputcreator
